move GoodsContext to index
